Refactored advanced analysis to tools package.
