+++ src/Tools/Average_Preprocessing/__init__.py
@@ -1,0 +1,9 @@
+# src/Tools/Average_Preprocessing/__init__.py
+
+from .advanced_analysis      import AdvancedAnalysisWindow
+from .advanced_analysis_core import run_advanced_averaging_processing
+
+__all__ = [
+    "AdvancedAnalysisWindow",
+    "run_advanced_averaging_processing",
+]
